Fix Bug 1.format time: (php->java) var time = SimpleDateFormat("YYYY-MM-dd HH:mm").format(content.date * 1000) 2.web view: override url loading

+++ app/src/main/java/io/github/jixiaoyong/todos/bean/ContentBean.kt
@@ -3,7 +3,7 @@ package io.github.jixiaoyong.todos.bean
 import com.google.gson.annotations.SerializedName
 
 data class ContentBean(@SerializedName("date")
-                       var date: Int = 0,
+                       var date: Long = 0,
                        @SerializedName("user_id")
                        var userId: Long = 0,
                        @SerializedName("tag")
