+++ app/src/main/java/io/github/jixiaoyong/todos/bean/Data.kt
@@ -3,7 +3,7 @@ package io.github.jixiaoyong.todos.bean
 import com.google.gson.annotations.SerializedName
 
 data class Data(@SerializedName("date")
-                val date: Int = 0,
+                val date: Long = 0,
                 @SerializedName("email")
                 val email: String? = null,
                 @SerializedName("username")
